+++ src/autopilot/calc/__init__.py
@@ -1,0 +1,5 @@
+"""Calculations."""
+
+from . import single
+
+__all__ = ["single"]